requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24.0

# Similarity Search
faiss-cpu>=1.8.0  # Use faiss-gpu if you have an NVIDIA card

# Image Processing & Utilities
Pillow>=10.0.0
tqdm>=4.66.0
pyyaml>=6.0.0
scikit-learn>=1.5.2

# Optional: For Web/API stage
fastapi
uvicorn